Revert "move formatPropertyValue to Models"
This reverts commit aa50298c3c67d7e770db1f4a616a655d54a4a9da.
