+++ lib/Core/formatPropertyValue.js
